requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-huggingface==0.0.3
langchain-groq==0.1.9
transformers==4.44.2
torch==2.4.0
tensorflow==2.17.0
huggingface-hub==0.24.6
sentence-transformers==3.0.1
faiss-cpu==1.8.0
streamlit==1.38.0
plotly==5.23.0
pandas==2.2.2
numpy==1.26.4
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
yfinance==0.2.43
groq==0.11.0
